Norm. Played around with ideas for ft_signal.c

+++ ft_signal.c
@@ -3,31 +3,39 @@
 #include <sys/types.h>
 #include <sys/wait.h>
 #include "minishell.h"
+#include <setjmp.h>
 
 pid_t childpid;
 char *buf;
 static int i;
+jmp_buf jump;
 
 void    ft_kill(int signum)
 {
 	i++;
-	if (childpid != 0)
+	if (childpid != 0 && childpid != 1)
 	{
 		kill(childpid, SIGTERM);
 	}
-	if (1 > 0)
-		ft_putstr("\b \b \b \b \b \b");
-	ft_putchar('\r');
-	// ft_putstr("\n$>");
+	if (childpid == 1)
+	{
+		ft_putendl("Wow");
+		longjmp(jump, 1);
+		ft_putchar('\n');
+	}
 }
 
 int main()
 {
     pid_t pid;
-
+	jmp_buf jump;
 	buf = NULL;
+
+	childpid = 1;
     signal(SIGINT, ft_kill);
-	while(1)
+	if (setjmp(jump))
+			goto LOOP;
+	LOOP: while(1)
 	{
 		buf = readline("Here we go");
 		if (!buf)
